Refactor tours package organization
