+++ lib/galaxy/tours/__init__.py
@@ -1,96 +1,7 @@
-"""
-This module manages loading/etc of Galaxy interactive tours.
-"""
-import logging
-import os
-
-import yaml
-from pydantic import parse_obj_as
-
-from galaxy import util
-from galaxy.tours import abc
-from galaxy.tours.schema import TourList
-
-
-log = logging.getLogger(__name__)
-
-
-def load_steps(contents_dict):
-    #  Some of this can be done on the clientside.  Maybe even should?
-    title_default = contents_dict.get('title_default', None)
-    for step in contents_dict['steps']:
-        if 'intro' in step:
-            step['content'] = step.pop('intro')
-        if 'position' in step:
-            step['placement'] = step.pop('position')
-        if 'element' not in step:
-            step['orphan'] = True
-        if title_default and 'title' not in step:
-            step['title'] = title_default
-
-
-@abc.ToursRegistry.register
-class ToursRegistry:
-    def __init__(self, tour_directories):
-        self.tour_directories = util.config_directories_from_setting(tour_directories)
-        self.load_tours()
-
-    def tours_by_id_with_description(self):
-        tours = []
-        for k in self.tours.keys():
-            tourdata = {
-                'id': k,
-                'description': self.tours[k].get('description'),
-                'name': self.tours[k].get('name'),
-                'tags': self.tours[k].get('tags')
-            }
-            tours.append(tourdata)
-        return parse_obj_as(TourList, tours)
-
-    def load_tour(self, tour_id):
-        for tour_dir in self.tour_directories:
-            tour_path = os.path.join(tour_dir, tour_id + ".yaml")
-            if not os.path.exists(tour_path):
-                tour_path = os.path.join(tour_dir, tour_id + ".yml")
-            if os.path.exists(tour_path):
-                return self._load_tour_from_path(tour_path)
-
-    def load_tours(self):
-        self.tours = {}
-        for tour_dir in self.tour_directories:
-            for filename in os.listdir(tour_dir):
-                if self._is_yaml(filename):
-                    self._load_tour_from_path(os.path.join(tour_dir, filename))
-        return self.tours_by_id_with_description()
-
-    def reload_tour(self, path):
-        # We may safely assume that the path is within the tour directory
-        filename = os.path.basename(path)
-        if self._is_yaml(filename):
-            self._load_tour_from_path(path)
-
-    def tour_contents(self, tour_id):
-        # Extra format translation could happen here (like the previous intro_to_tour)
-        # For now just return the loaded contents.
-        return self.tours.get(tour_id, None)
-
-    def _is_yaml(self, filename):
-        return filename.endswith('.yaml') or filename.endswith('.yml')
-
-    def _load_tour_from_path(self, tour_path):
-        filename = os.path.basename(tour_path)
-        tour_id = os.path.splitext(filename)[0]
-        try:
-            with open(tour_path) as handle:
-                tour = yaml.safe_load(handle)
-                load_steps(tour)
-                self.tours[tour_id] = tour
-                log.info("Loaded tour '%s'" % tour_id)
-                return tour
-        except OSError:
-            log.exception("Tour '%s' could not be loaded, error reading file.", tour_id)
-        except yaml.error.YAMLError:
-            log.exception("Tour '%s' could not be loaded, error within file.  Please check your yaml syntax.", tour_id)
-        except TypeError:
-            log.exception("Tour '%s' could not be loaded, error within file. Possibly spacing related. Please check your yaml syntax.", tour_id)
-        return None
+from ._impl import build_tours_registry  # noqa
+from ._interface import ToursRegistry  # noqa
+from ._schema import Tour  # noqa
+from ._schema import TourCore  # noqa
+from ._schema import TourStep  # noqa
+from ._schema import TourDetails  # noqa
+from ._schema import TourList  # noqa
